Refactored provider code.
