+++ src/provider/utils/fetchAndPrepareOpenApiData.ts
